Cherrypicked from UE5/Dev-TensorWorks
Fix PixelStreaming NonUnity Linux

#rb luke.bermingham
#fyi william.belcher, aidan.possemiers, mattias.jansson
#jira UE-159187
#preflight 62cd1266ad490b52daf8a2e6

[CL 21053344 by Nick Pace in ue5-main branch]

+++ Engine/Plugins/Media/PixelStreaming/Source/PixelStreaming/Private/FrameAdapterProcessRHIToH264.cpp
@@ -40,7 +40,7 @@ namespace UE::PixelStreaming
 
 	void FFrameAdapterProcessRHIToH264::BeginProcess(const IPixelStreamingInputFrame& InputFrame, TSharedPtr<IPixelStreamingAdaptedOutputFrame> OutputBuffer)
 	{
-		checkf(InputFrame.GetType() == EPixelStreamingInputFrameType::RHI, TEXT("Incorrect source frame coming into frame adapter process."));
+		checkf(InputFrame.GetType() == static_cast<int32>(EPixelStreamingInputFrameType::RHI), TEXT("Incorrect source frame coming into frame adapter process."));
 
 		FRHICommandListImmediate& RHICmdList = FRHICommandListExecutor::GetImmediateCommandList();
 		RHICmdList.EnqueueLambda([this](FRHICommandListImmediate&) { MarkAdaptProcessStarted(); });
